refactor(sources): split stac_resolve into taxonomy, identity, asset gate and by-search
stac_resolve.py crossed 1013 lines with its cap raised twice. Split along
its natural seams into sibling modules: stac_resolve_taxonomy.py (the
StacResolution dataclass and closed verdict vocabulary), stac_resolve_
identity.py (URL/body identity derivation shared by every path),
stac_resolve_asset_gate.py (turns one fetched item document into a
resolution, including the SSRF + COG probe), and stac_resolve_by_search.py
(the by-search fallback). stac_resolve.py keeps resolve_stac_binding (the
by-URL entry point) and re-exports everything else so every existing
import path keeps working.

No behavior change: every moved function is byte-identical, and the
mock-patch targets in test_stac_refresh_1266.py / test_refresh_gate_1269.py
that reached into fetch_cog_info/validate_url_for_ssrf now point at
stac_resolve_asset_gate, the module that actually calls them post-split.
